+++ src/utils.js
@@ -1,0 +1,3 @@
+import readlineSync from 'readline-sync'
+
+export const ask = askStr => readlineSync.question(`${askStr} \n>`)
